# Added a specialized version of `Deferred` based on `IOFiber`'s machinery