--- a/benchmarks/src/main/scala/cats/effect/benchmarks/QueueBenchmark.scala
+++ b/benchmarks/src/main/scala/cats/effect/benchmarks/QueueBenchmark.scala
@@ -33,7 +33,7 @@ import java.util.concurrent.TimeUnit
  *
  * Or to run the benchmark from within sbt:
  *
- * jmh:run -i 10 -wi 10 -f 2 -t 1 cats.effect.benchmarks.QueueBenchmark
+ * Jmh / run -i 10 -wi 10 -f 2 -t 1 cats.effect.benchmarks.QueueBenchmark
  *
  * Which means "10 iterations", "10 warm-up iterations", "2 forks", "1 thread". Please note that
  * benchmarks should be usually executed at least in 10 iterations (as a rule of thumb), but