--- a/core/shared/src/main/scala/cats/effect/IO.scala
+++ b/core/shared/src/main/scala/cats/effect/IO.scala
@@ -802,9 +802,6 @@ sealed abstract class IO[+A] private () extends IOPlatform[A] {
   def void: IO[Unit] =
     map(_ => ())
 
-  def voidError(implicit ev: A <:< Unit): IO[Unit] =
-    asInstanceOf[IO[Unit]].handleError(_ => ())
-
   /**
    * Converts the source `IO` into any `F` type that implements the [[LiftIO]] type class.
    */