--- a/core/shared/src/main/scala/cats/effect/IODeferred.scala
+++ b/core/shared/src/main/scala/cats/effect/IODeferred.scala
@@ -77,6 +77,6 @@ private final class IODeferred[A] extends Deferred[IO, A] {
   }
 }
 
-object IODeferred {
+private object IODeferred {
   private val Sentinel = new AnyRef
 }