--- a/tests/shared/src/test/scala/cats/effect/IOSpec.scala
+++ b/tests/shared/src/test/scala/cats/effect/IOSpec.scala
@@ -1584,6 +1584,12 @@ class IOSpec extends BaseSpec with Discipline with IOPlatformSpecification {
         implicitly)
     }
 
+    "produce a specialized version of Deferred" in real {
+      IO.deferred[Unit] flatMap { d =>
+        IO(d must haveClass[IODeferred[_]])
+      }
+    }
+
     platformSpecs
   }