Experimental reimplementation of `Deferred` based on `IOFiber`'s internal machinery rather than the more generic `LongMap` approach. Benchmarks incoming. One unfortunate discovery as part of working on this is the fact that the `Deferred` API is… distressingly open. `unsafe` is quite annoying, but `AsyncDeferred` is even worse. May have to put more thought into fixing this.

Fixes #3271