@djspiewak 2022-11-26
Performance results are in! Looks like this is around 20-40% faster than the original. The higher count tests are more accurate than the lower count ones since they factor out the overhead a bit better. The `complete` benchmark is… bad… and I need to find a better way to do it. Unfortunately, this one is sort of fundamentally a race condition, which I suspect is part of why we don't see a major improvement in this scenario (~5%), and also why one of the results is wildly out of step with the others.

## Before

```
[info] Benchmark                   (count)   Mode  Cnt     Score    Error  Units
[info] DeferredBenchmark.cancel         10  thrpt   10    90.653 ±  1.991  ops/s
[info] DeferredBenchmark.cancel        100  thrpt   10    10.547 ±  0.188  ops/s
[info] DeferredBenchmark.cancel       1000  thrpt   10     0.834 ±  0.039  ops/s
[info] DeferredBenchmark.complete       10  thrpt   10    62.500 ±  0.385  ops/s
[info] DeferredBenchmark.complete      100  thrpt   10     3.766 ±  0.038  ops/s
[info] DeferredBenchmark.complete     1000  thrpt   10     2.743 ±  0.041  ops/s
[info] DeferredBenchmark.get            10  thrpt   10  3369.843 ± 20.579  ops/s
[info] DeferredBenchmark.get           100  thrpt   10   533.762 ±  2.963  ops/s
[info] DeferredBenchmark.get          1000  thrpt   10    56.230 ±  0.297  ops/s
```

## After

```
[info] Benchmark                   (count)   Mode  Cnt     Score    Error  Units
[info] DeferredBenchmark.cancel         10  thrpt   10    97.689 ±  1.219  ops/s
[info] DeferredBenchmark.cancel        100  thrpt   10    12.487 ±  0.292  ops/s
[info] DeferredBenchmark.cancel       1000  thrpt   10     1.077 ±  0.053  ops/s
[info] DeferredBenchmark.complete       10  thrpt   10    65.497 ±  0.245  ops/s
[info] DeferredBenchmark.complete      100  thrpt   10    15.663 ±  0.897  ops/s
[info] DeferredBenchmark.complete     1000  thrpt   10     2.833 ±  0.046  ops/s
[info] DeferredBenchmark.get            10  thrpt   10  4019.782 ± 13.285  ops/s
[info] DeferredBenchmark.get           100  thrpt   10   743.074 ±  3.311  ops/s
[info] DeferredBenchmark.get          1000  thrpt   10    80.084 ±  0.312  ops/s
```