@armanbilge 2022-12-04
> The `complete` benchmark is… bad… and I need to find a better way to do it.

Sorry, why is it bad exactly?