@filipwiech 2022-12-04
Some ideas for further optimizations based on @armanbilge's recent work (#3281 & #3300):

- Try to avoid wrapping `AtomicReference`.
- Specialize `Deferred` for `Boolean`, `Int`, `Long`.

Not sure how much sense they make and most likely you've already thought of them, but wanted to mention it just in case. 🙂