@armanbilge 2022-12-04
> * Try to avoid wrapping `AtomicReference`.

I tried this as part of https://github.com/typelevel/cats-effect/pull/3281 using a `AtomicReferenceFieldUpdater` but unfortunately it had worse performance, see https://github.com/typelevel/cats-effect/pull/3281#issuecomment-1327055441.