@filipwiech 2022-12-04
You're right, the idea might indeed be not-so-great. Although it could be interesting to benchmark the impact on memory allocation. There's also `AtomicIntegerFieldUpdater` and `AtomicLongFieldUpdater` that may be of interest in the specialized cases (there isn't one for `Boolean`?), but I admit that it looks like a long shot, sorry for the noise. 👍