@armanbilge 2022-12-04
Yeah, it's a memory / performance trade-off. But since most refs are get/set/update many more times than they are allocated, probably better to bias towards performance.

The specializations are also interesting, except my other PR also found no performance improvement from doing that 😂 see https://github.com/typelevel/cats-effect/pull/3300#issuecomment-1336407466.

Optimization is hard :) @filipwiech it's hard to know what's a good idea or not until you try it and benchmark it.