@djspiewak 2022-12-24
@armanbilge 

> Sorry, why is it bad exactly?

Sorry didn't reply to this. The benchmark is as follows:

```scala
  @Benchmark
  def complete(): Unit = {
    IO.deferred[Unit]
      .flatMap { d => d.get.parReplicateA_(count) &> d.complete(()) }
      .replicateA_(1000)
      .unsafeRunSync()
  }
```

So the problem here is the fact that it just reduces to the `get` benchmark very, very quickly:

```scala
  @Benchmark
  def get(): Unit = {
    IO.deferred[Unit]
      .flatMap(d => d.complete(()) *> d.get.replicateA_(count))
      .replicateA_(1000)
      .unsafeRunSync()
  }
```

Notice the only difference here is the race condition on `complete`'s scheduling!

What I'm *trying* to measure is the cost when a large number of fibers are enqueued on `get` and then released simultaneously by a single `complete`. The problem is that Cats Effect's concurrency model makes this a very difficult thing to model deterministically, which in turn makes it immensely hard to measure from a performance standpoint.