@armanbilge 2022-12-24
> The problem is that Cats Effect's concurrency model makes this a very difficult thing to model deterministically, which in turn makes it immensely hard to measure from a performance standpoint.

Ah, indeed! Thanks for the explanation :)