@djspiewak 2022-12-24
> Now that CallbackStack has become even more important I'd like to follow-up with a specialized implementation for JS.

I'm curious if you can actually beat the performance on JS. From my understanding of how `AtomicReference` compiles, it should be effectively equivalent to the best you can do even taking advantage of the single-threaded runtime.