@armanbilge 2022-12-24
> I'm curious if you can actually beat the performance on JS. From my understanding of how `AtomicReference` compiles, it should be effectively equivalent to the best you can do even taking advantage of the single-threaded runtime.

Really? Isn't this basically a thread-safe linked list? On JS couldn't we just use a native stack aka `js.Array` ?