@djspiewak 2022-12-24
Oh right it's a stack not a queue. Yeah you could totally beat it.